+++ common/src/main/java/net/xolt/freecam/config/gui/AutoConfigExtensions.java
@@ -22,6 +22,7 @@ public class AutoConfigExtensions {
 
     public static void apply(Class<? extends ConfigData> configClass) {
         GuiRegistry registry = AutoConfig.getGuiRegistry(configClass);
+        ModBindingsConfigImpl.apply(registry);
         VariantTooltipImpl.apply(registry);
         BoundedContinuousImpl.apply(registry);
     }